refactor: extract flattenSchema utility
